requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docstring_parser==0.16
entrypoints==0.4
Faker==28.0.0
favicon==0.7.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gcsfs==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.11.10
google-cloud-firestore==2.17.2
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.17.1
google-cloud-vision==3.7.4
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.62.3
htbuilder==0.6.2
idna==3.8
imagecodecs==2024.6.1
imageio==2.35.1
importlib-metadata==6.11.0
importlib_resources==6.4.4
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.4.0
multidict==6.0.5
narwhals==1.5.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==23.2
pandas==2.2.2
Pillow==9.5.0
plotly==5.23.0
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.9
Pympler==1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
soupsieve==2.6
st-annotated-text==4.0.1
st-pages==0.5.0
st-theme==1.2.2
stqdm==0.0.5
streamlit==1.40.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-image-comparison==0.0.4
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
tifffile==2024.8.24
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.2
validators==0.33.0
yarl==1.9.4
zipp==3.20.0
